# services/sdk-ts/openapi.yaml
openapi: 3.1.0
info:
  title: deployserver operator API
  version: 1.0.0
  description: >
    The operator-facing REST API for deployserver. All endpoints under
    `/api/v1` require a bearer OIDC ID token (obtain one with
    `deployctl auth login`); in dev mode (no OIDC configured) auth is
    open. Permissions are enforced per-endpoint via RBAC — the required
    permission is noted in each operation's description. This document is
    kept honest by a contract test that asserts the running routes and
    these paths are in exact correspondence.
  license:
    name: Apache-2.0
servers:
  - url: https://deploy.example.com
    description: Replace with your DEPLOY_FQDN.
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/machines:
    get:
      tags: [machines]
      summary: List machines
      description: "Requires `machine.read`."
      responses:
        "200":
          description: Machines
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Machine" } }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [machines]
      summary: Register a machine
      description: "Requires `machine.write`."
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMachine" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Machine" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/machines/{id}:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [machines]
      summary: Get a machine
      description: "Requires `machine.read`."
      responses:
        "200": { description: Machine, content: { application/json: { schema: { $ref: "#/components/schemas/Machine" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [machines]
      summary: Update a machine (partial)
      description: "Requires `machine.write`. Omitted fields are unchanged; an empty `default_profile_id` clears it."
      requestBody:
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateMachine" } } }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/Machine" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [machines]
      summary: Delete a machine
      description: "Requires `machine.write`."
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/machines/{id}/wake:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [machines]
      summary: Wake-request history
      description: "Requires `machine.read`."
      responses:
        "200": { description: Wake requests, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/WakeRequest" } } } } }
    post:
      tags: [machines]
      summary: Queue a wake-on-LAN
      description: "Requires `deployment.create`. Site defaults to the machine's `attributes.site`."
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                at: { type: string, format: date-time, description: "RFC3339; empty = next agent poll" }
                site: { type: string }
      responses:
        "202": { description: Queued, content: { application/json: { schema: { type: object, properties: { wake_id: { type: string }, site: { type: string } } } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/profiles:
    get:
      tags: [profiles]
      summary: List deployment profiles
      description: "Requires `image.read`."
      responses:
        "200": { description: Profiles, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Profile" } } } } }
    post:
      tags: [profiles]
      summary: Create a profile
      description: "Requires `image.write`."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, image_id]
              properties:
                name: { type: string }
                image_id: { type: string, format: uuid }
                answer_file_vars: { type: object, additionalProperties: true }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Profile" } } } }
  /api/v1/profiles/{id}:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [profiles]
      summary: Get a profile with its answer-file templates
      description: "Requires `image.read`."
      responses:
        "200": { description: Profile }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [profiles]
      summary: Update a profile
      description: "Requires `image.write`."
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses:
        "200": { description: Updated }
    delete:
      tags: [profiles]
      summary: Delete a profile
      description: "Requires `image.write`."
      responses:
        "204": { description: Deleted }
  /api/v1/profiles/{id}/templates:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    put:
      tags: [profiles]
      summary: Upsert an answer-file template
      description: "Requires `image.write`. `kind` is one of autoinstall|kickstart|preseed|cloud-init|ignition|unattend."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, body]
              properties:
                kind: { type: string }
                body: { type: string }
      responses:
        "200": { description: Upserted }
  /api/v1/profiles/{id}/templates/{kind}:
    parameters:
      - { $ref: "#/components/parameters/IdPath" }
      - { name: kind, in: path, required: true, schema: { type: string } }
    delete:
      tags: [profiles]
      summary: Delete an answer-file template
      description: "Requires `image.write`."
      responses:
        "204": { description: Deleted }
  /api/v1/images:
    get:
      tags: [images]
      summary: List images
      description: "Requires `image.read`."
      responses:
        "200": { description: Images, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Image" } } } } }
    post:
      tags: [images]
      summary: Create an image
      description: "Requires `image.write`."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, os_family, os_version, arch]
              properties:
                name: { type: string }
                os_family: { type: string, enum: [windows, linux] }
                os_version: { type: string }
                arch: { type: string, enum: [amd64, arm64] }
                description: { type: string }
                media: { $ref: "#/components/schemas/ImageMedia" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Image" } } } }
  /api/v1/images/{id}:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [images]
      summary: Get an image
      description: "Requires `image.read`."
      responses: { "200": { description: Image, content: { application/json: { schema: { $ref: "#/components/schemas/Image" } } } } }
    patch:
      tags: [images]
      summary: Update an image (name/description/media)
      description: "Requires `image.write`."
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses: { "200": { description: Updated } }
    delete:
      tags: [images]
      summary: Delete an image
      description: "Requires `image.write`."
      responses: { "204": { description: Deleted } }
  /api/v1/images/{id}/versions:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [images]
      summary: List image versions
      description: "Requires `image.read`."
      responses: { "200": { description: Versions } }
    post:
      tags: [images]
      summary: Link an uploaded blob as a new image version
      description: "Requires `image.write`. Register the blob with POST /api/v1/blobs first."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [blob_id]
              properties: { blob_id: { type: string, format: uuid }, version_tag: { type: string } }
      responses: { "201": { description: Version created } }
  /api/v1/blobs:
    post:
      tags: [images]
      summary: Register a blob and get a presigned upload URL
      description: "Requires `image.write`. Upload the bytes to the returned URL (PUT), then link via image or driver-pack version."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sha256, size_bytes]
              properties:
                sha256: { type: string, description: "64 lowercase hex chars" }
                size_bytes: { type: integer, format: int64 }
                filename: { type: string }
                role: { type: string, enum: [images, drivers, blobs] }
      responses:
        "201":
          description: Presigned upload
          content:
            application/json:
              schema:
                type: object
                properties:
                  blob_id: { type: string, format: uuid }
                  bucket: { type: string }
                  key: { type: string }
                  upload_url: { type: string }
                  expires_in_seconds: { type: integer }
        "503": { description: Object store not configured }
  /api/v1/driver-packs:
    get:
      tags: [driver-packs]
      summary: List driver-pack versions
      description: "Requires `driverpack.read` or `image.read`."
      responses: { "200": { description: Driver packs } }
    post:
      tags: [driver-packs]
      summary: Create a driver-pack version with match rules
      description: "Requires `image.write`. Register the .zip blob (role=drivers) first."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vendor, model, os_family, blob_id, rules]
              properties:
                vendor: { type: string }
                model: { type: string }
                os_family: { type: string, enum: [windows, linux] }
                os_version: { type: string }
                version_tag: { type: string }
                blob_id: { type: string, format: uuid }
                rules:
                  type: array
                  items:
                    type: object
                    properties:
                      type: { type: string, enum: [pci-vid-did, dmi-product, dmi-baseboard, dmi-vendor, os-version] }
                      value: { type: string }
      responses: { "201": { description: Created } }
  /api/v1/driver-packs/versions/{id}:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    delete:
      tags: [driver-packs]
      summary: Delete a driver-pack version
      description: "Requires `image.write`."
      responses: { "204": { description: Deleted }, "404": { $ref: "#/components/responses/NotFound" } }
  /api/v1/vendor-driverpacks:
    get:
      tags: [driver-packs]
      summary: Search the vendor driver-pack catalogs
      description: "Requires `driverpack.read`. Searches Lenovo's published enterprise driver-pack catalog by model name or machine type. Empty query returns nothing."
      parameters:
        - { name: q, in: query, schema: { type: string }, description: "Model name or machine type, e.g. 'X1 Carbon' or '20XW'" }
      responses:
        "200": { description: Matching catalog entries }
        "502": { description: Vendor catalog unreachable and no cached copy }
  /api/v1/vendor-driverpacks/fetch:
    post:
      tags: [driver-packs]
      summary: Queue a vendor driver-pack download
      description: "Requires `driverpack.write`. The URL must be one returned by the search endpoint — the server only downloads addresses read from the vendor's own catalog. The pack is streamed into the blob store, sha256-verified against the catalog, and registered with dmi-product-prefix match rules for its machine types."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string }
      responses:
        "202": { description: Fetch job queued }
        "400": { description: URL not present in the vendor catalog }
        "503": { description: Blob store not configured }
  /api/v1/vendor-driverpacks/jobs:
    get:
      tags: [driver-packs]
      summary: List vendor fetch jobs
      description: "Requires `driverpack.read`. Most recent first."
      parameters:
        - { name: limit, in: query, schema: { type: integer } }
      responses: { "200": { description: Fetch jobs with state and error detail } }
...
  /api/v1/catalog/install:
    post:
      tags: [catalog]
      summary: Install an image from the catalog
      description: "Requires `image.write`."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entry_id]
              properties: { entry_id: { type: string }, name: { type: string } }
      responses: { "201": { description: Image created from catalog } }
  /api/v1/jobs:
    get:
      tags: [jobs]
      summary: List deployment jobs
      description: "Requires `job.read`. Filter with `state`, `machine`, `limit`."
      parameters:
        - { name: state, in: query, schema: { type: string } }
        - { name: machine, in: query, schema: { type: string, format: uuid } }
        - { name: limit, in: query, schema: { type: integer } }
      responses: { "200": { description: Jobs, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Job" } } } } } }
  /api/v1/jobs/{id}:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [jobs]
      summary: Get a job with its event trail
      description: "Requires `job.read`."
      responses: { "200": { description: Job + events }, "404": { $ref: "#/components/responses/NotFound" } }
  /api/v1/jobs/{id}/cancel:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    post:
      tags: [jobs]
      summary: Cancel a non-terminal job
      description: "Requires `deployment.create`. Revokes the job's boot tokens."
      responses: { "204": { description: Cancelled }, "409": { $ref: "#/components/responses/Conflict" } }
  /api/v1/deployments/issue:
    post:
      tags: [deployments]
      summary: Issue a deployment code
      description: "Requires `deployment.create` (capture additionally needs `image.write`)."
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IssueDeployment" }
      responses:
        "200":
          description: Code issued
          content:
            application/json:
              schema: { type: object, properties: { code: { type: string }, expires_at: { type: string, format: date-time } } }
  /api/v1/deployments/bulk:
    post:
      tags: [deployments]
      summary: Issue deployment codes for many machines
      description: "Requires `deployment.create`. Up to 100 machines; always 200 with per-machine results."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [machine_ids, profile_id]
              properties:
                machine_ids: { type: array, items: { type: string, format: uuid } }
                profile_id: { type: string, format: uuid }
                ttl_seconds: { type: integer }
                wake: { type: boolean }
      responses:
        "200":
          description: Per-machine results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        machine_id: { type: string }
                        code: { type: string }
                        expires_at: { type: string }
                        wake_id: { type: string }
                        error: { type: string }
  /api/v1/reports/summary:
    get:
      tags: [reports]
      summary: Deployment summary for a window
      description: "Requires `job.read`. `since` accepts 24h/7d/30d style values."
      parameters: [{ $ref: "#/components/parameters/Since" }]
      responses: { "200": { description: Summary, content: { application/json: { schema: { $ref: "#/components/schemas/ReportSummary" } } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest: { description: Invalid request }
    Forbidden: { description: Missing the required permission }
    NotFound: { description: Not found }
    Conflict: { description: State conflict }
  schemas:
    Machine:
      type: object
      properties:
        ID: { type: string, format: uuid }
        AssetTag: { type: [string, "null"] }
        MACPrimary: { type: [string, "null"] }
        Vendor: { type: [string, "null"] }
        Model: { type: [string, "null"] }
        DefaultProfileID: { type: [string, "null"], format: uuid }
        Attributes: { type: object, additionalProperties: true }
        CreatedAt: { type: string, format: date-time }
...